% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDataSSP.R
\name{chave_list}
\alias{chave_list}
\title{Index para parear dados}
\usage{
chave_list(x, vec = seq(645))
}
\arguments{
\item{x}{e um vetor qualquer, precisamos do tamanho dele}

\item{vec}{e o vetor de indexacao (1:645 municipios) que sera replicado `x` vezes}
}
\value{
uma `list` com dois vetores igual a `vec`.
}
\description{
Index para parear dados
}
