% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDataSSP.R
\name{download_table_sp}
\alias{download_table_sp}
\title{Extract data from SSP}
\usage{
download_table_sp(ano, municipio, type = c("ctl00$conteudo$btnPolicial",
  "ctl00$conteudo$btnMensal"))
}
\arguments{
\item{ano}{is a number between 2002 e 2018 indicating the year,}

\item{municipio}{is a number between 1 and 645 indicating the municipality (organized by alphabetical
order)}

\item{type}{a character string indicating the data resquested to the SSP's website. Use 
\code{"ctl00$conteudo$btnMensal"} if you want to extract criminal ocurrencies or
\code{"ctl00$conteudo$btnPolicial"} to request police indicators.}
}
\value{
a tibble
}
\description{
Extract data avaliable in SSP's website: (1) number of criminal ocurrencies (2) police productivity indicators,
by month and by municipalities.
}
\references{
Web scraping do site da Secretaria de Seguranca Publica de Sao Paulo.
\link[www.curso-r.com/blog/2017/05/19/2017-05-19-scrapper-ssp/]{Blog do Curso R}. 2017. ultimo acesso em: 11/07/2018

\link[www.ssp.sp.gov.br/Estatistica/Pesquisa.aspx]{Secretaria de Seguranca Publica do estado de Sao Paulo}
}
