% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDataSSP.R
\name{municipio}
\alias{municipio}
\title{Get city index by it's name}
\usage{
municipio(x)
}
\arguments{
\item{x}{a character string with the city name e.g. "SÃO PAULO"}
}
\value{
the index number of the city
}
\description{
Function to find the city \emph[index] in Sao Paulo state.
}
\details{
Capital letters only, separated by spaces and without accents
}
