% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDataSSP.R
\name{\%>\%}
\alias{\%>\%}
\alias{municipio}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

municipio(x)
}
\arguments{
\item{x}{a character string with the city name e.g. "SÃO PAULO"}
}
\value{
the index number of the city

#' @importFrom magrittr %>%
}
\description{
See \code{\link[magrittr]{\%>\%}} for more details.

Function to find the city \emph[index] in Sao Paulo state.
}
\details{
Capital letters only, separated by spaces and without accents
}
\keyword{internal}
