% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeadeData.R
\docType{data}
\name{seade_pop_state}
\alias{seade_pop_state}
\title{Total population count by year and age groups}
\format{A \code{tibble} with 1,248 rows and 5 variables:
\describe{
\item{\code{cod}}{Variable name in \code{seade_data} object}
\item{\code{populacao}}{Integer, population count in Sao Paulo state}
\item{\code{genero}}{Female or Male}
\item{\code{faixa_etaria}}{Age groups}
\item{\code{ano}}{Year}
}}
\usage{
seade_pop_state
}
\description{
This dataset provides total population in Sao Paulo state by year and age groups.
}
\keyword{datasets}
