% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeadeData.R
\docType{data}
\name{seade_pop_variables_names}
\alias{seade_pop_variables_names}
\title{Age groups and native variables id's on \code{seade_data}.}
\format{A \code{data.frame} with 54 rows and 2 variables:
\describe{
  \item{\code{descricao}}{The variable name in \code{seade_pop} obj.}
  \item{\code{var_id}}{The native ID in the raw data}
}}
\source{
\url{https://www.seade.gov.br/}
}
\usage{
seade_pop_variables_names
}
\description{
With this function you can quicly reviwe the age groups and the variables names on \code{seade_data} object
}
\keyword{datasets}
