% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SspData.R
\docType{data}
\name{ssp_ocorrencias_crimes}
\alias{ssp_ocorrencias_crimes}
\title{Crime counts in Sao Paulo state (2002 - 2017).}
\format{A \code{tibble} with 8,385 rows and 16 variables.}
\source{
\url{http://www.ssp.sp.gov.br/Estatistica/Pesquisa.aspx}.
}
\usage{
ssp_ocorrencias_crimes
}
\description{
A dataset containing crime counts by municipalities in Sao paulo state, Brazil.
}
\seealso{
Use \code{\link{ssp_variable_names}} to see details about this dataset.
}
\keyword{datasets}
