% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SspData.R
\docType{data}
\name{ssp_variables_names}
\alias{ssp_variables_names}
\title{Variable names in SSP datasets.}
\format{A \code{tibble} with 32 rows and 3 variables.
\describe{
  \item{\code{variaveis_tidy}}{Variables names in tidy format, to build an R friendly dataset}
  \item{\code{variable_raw}}{Original variable names (as you can see in SSP's website)}
  \item{\code{dataset}}{Character string indicating the dataset in \code{spcrimr}}
  }}
\source{
\url{http://www.ssp.sp.gov.br/Estatistica/Pesquisa.aspx}.
}
\usage{
ssp_variables_names
}
\description{
A dataset containing variables names in tidy format and raw (exactly the same in SSP's website)
This dataset contain details about the variable names in raw data.
}
\keyword{datasets}
