% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeadeData.R
\docType{data}
\name{seade_pop}
\alias{seade_pop}
\title{Population count in Sao Paulo state (1980 - 2018).}
\format{A \code{tibble} with 25.155 rows and 39 variables:
\describe{
  \item{Localidades (\code{municipio})}{cities in Sao Paulo}
  \item{Periodos (\code{ano})}{year (1980 - 2018)}
  \item{Populacao (\code{pop})}{Total opulation count}
  \item{Populacao Feminina (+ages, \code{pop_fem})}{Female population count}
  \item{Populacao Masculina (+ages, \code{pop_msc})}{Male population count}
  \item{Cod. IBGE (\code{cd_geocmu})}{Geo code}
}}
\source{
\url{https://www.seade.gov.br/}
}
\usage{
seade_pop
}
\description{
A dataset containing resident population in Sao Paulo state by city (645) sex and age. The data was collected in 
Fundacao SEADE's website.
}
\keyword{datasets}
